requirements.txt
autopep8==1.3.3
bytecode==0.5
neo-boa==0.3.7
pycodestyle==2.3.1
logzero==1.5.0